+++ xarray/core/nputils.py
@@ -1,9 +1,10 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
+from __future__ import absolute_import, division, print_function
+
+import warnings
+
 import numpy as np
 import pandas as pd
-import warnings
+
 from . import npcompat