+++ xarray/core/ops.py
@@ -5,18 +5,15 @@ NumPy's __array_ufunc__ and mixin classes instead of the unintuitive "inject"
 functions.
 """
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
+from __future__ import absolute_import, division, print_function
 
 import operator
 
 import numpy as np
 
-from . import dtypes
-from . import duck_array_ops
-from .pycompat import PY3
+from . import dtypes, duck_array_ops
 from .nputils import array_eq, array_ne
+from .pycompat import PY3
 
 try:
     import bottleneck as bn