+++ xarray/tests/test_variable.py
@@ -1,35 +1,32 @@
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
+from __future__ import absolute_import, division, print_function
+
 from collections import namedtuple
 from copy import copy, deepcopy
 from datetime import datetime, timedelta
-from textwrap import dedent
-import pytest
-
 from distutils.version import LooseVersion
-import numpy as np
-import pytz
-import pandas as pd
+from textwrap import dedent
 
-from xarray import Variable, IndexVariable, Coordinate, Dataset
+import numpy as np
+import pandas as pd
+import pytest
+import pytz
+
+from xarray import Coordinate, Dataset, IndexVariable, Variable
 from xarray.core import indexing
-from xarray.core.variable import as_variable, as_compatible_data
-from xarray.core.indexing import (PandasIndexAdapter, LazilyIndexedArray,
-                                  BasicIndexer, OuterIndexer,
-                                  VectorizedIndexer, NumpyIndexingAdapter,
-                                  CopyOnWriteArray, MemoryCachedArray,
-                                  DaskIndexingAdapter)
+from xarray.core.common import full_like, ones_like, zeros_like
+from xarray.core.indexing import (
+    BasicIndexer, CopyOnWriteArray, DaskIndexingAdapter, LazilyIndexedArray,
+    MemoryCachedArray, NumpyIndexingAdapter, OuterIndexer, PandasIndexAdapter,
+    VectorizedIndexer)
 from xarray.core.pycompat import PY3, OrderedDict
-from xarray.core.common import full_like, zeros_like, ones_like
 from xarray.core.utils import NDArrayMixin
+from xarray.core.variable import as_compatible_data, as_variable
+from xarray.tests import requires_bottleneck
 
 from . import (
-    TestCase, source_ndarray, requires_dask, raises_regex, assert_identical,
-    assert_array_equal, assert_equal, assert_allclose)
-
-from xarray.tests import requires_bottleneck
+    TestCase, assert_allclose, assert_array_equal, assert_equal,
+    assert_identical, raises_regex, requires_dask, source_ndarray)
 
 
 class VariableSubclassTestCases(object):
@@ -1764,6 +1761,13 @@ class TestIndexVariable(TestCase, VariableSubclassTestCases):
         with pytest.warns(Warning, match='deprecated'):
             x = Coordinate('x', [1, 2, 3])
         assert isinstance(x, IndexVariable)
+
+    def test_datetime64(self):
+        # GH:1932  Make sure indexing keeps precision
+        t = np.array([1518418799999986560, 1518418799999996560],
+                     dtype='datetime64[ns]')
+        v = IndexVariable('t', t)
+        assert v[0].data == t[0]
 
     # These tests make use of multi-dimensional variables, which are not valid
     # IndexVariable objects: